Give meaningful names to decorated functions
This:
* Names decorated functions simply "memoized" in the case the input function is anonymous
* Names decorated functions "memoized(foo)" in the case the input function has a `name` of "foo"
* Names decorated functions "memoized(foo)" in the case the input function has a `displayName` of "foo", despite having a `name` of "bar".

It also gives a name to the default export of `memoize-one` of "memoizeOne", as it was previously "index" in every case but the minified version, where it was the empty string and will continue to be.

Upside: Makes for debugging these decorated functions much easier in debugging tools and stack traces.

Downside: more checks and more names add to the size of the compiled output.

Test Plan: `yarn test`. Import the `main` entry from a repl and confirm the export is named `memoizeOne`.